+++ tests/Feature/OrganizationTest.php
@@ -2,13 +2,12 @@
 
 namespace Tests\Feature;
 
+use App\Models\Invitation;
 use App\Models\Membership;
 use App\Models\Organization;
-use App\Models\Invitation;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\URL;
-
 use Illuminate\Support\Str;
 use Tests\TestCase;
 
@@ -18,18 +17,22 @@ class OrganizationTest extends TestCase
 
     public function test_users_can_create_organizations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
 
-        $response = $this->actingAs($user)->get('/en/organizations/create');
+        $response = $this->actingAs($user)->get(localized_route('organizations.create'));
         $response->assertStatus(200);
 
-        $response = $this->actingAs($user)->post('/en/organizations/create', [
+        $response = $this->actingAs($user)->post(localized_route('organizations.create'), [
             'name' => $user->name . ' Consulting',
             'locality' => 'Truro',
-            'region' => 'ns'
+            'region' => 'NS',
         ]);
 
-        $url = '/en/organizations/' . Str::slug($user->name . ' Consulting');
+        $url = localized_route('organizations.show', ['organization' => Str::slug($user->name . ' Consulting')]);
 
         $response->assertSessionHasNoErrors();
 
@@ -38,6 +41,10 @@ class OrganizationTest extends TestCase
 
     public function test_users_with_admin_role_can_edit_organizations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $organization = Organization::factory()
             ->hasAttached($user, ['role' => 'admin'])
@@ -49,13 +56,17 @@ class OrganizationTest extends TestCase
         $response = $this->actingAs($user)->put(localized_route('organizations.update', $organization), [
             'name' => $organization->name,
             'locality' => 'St John\'s',
-            'region' => 'nl'
+            'region' => 'NL',
         ]);
         $response->assertRedirect(localized_route('organizations.show', $organization));
     }
 
     public function test_users_without_admin_role_can_not_edit_organizations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $organization = Organization::factory()
             ->hasAttached($user, ['role' => 'member'])
@@ -67,13 +78,17 @@ class OrganizationTest extends TestCase
         $response = $this->actingAs($user)->put(localized_route('organizations.update', $organization), [
             'name' => $organization->name,
             'locality' => 'St John\'s',
-            'region' => 'nl'
+            'region' => 'NL',
         ]);
         $response->assertStatus(403);
     }
 
     public function test_non_members_can_not_edit_organizations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $other_user = User::factory()->create();
 
@@ -91,13 +106,17 @@ class OrganizationTest extends TestCase
         $response = $this->actingAs($user)->put(localized_route('organizations.update', $other_organization), [
             'name' => $other_organization->name,
             'locality' => 'St John\'s',
-            'region' => 'nl'
+            'region' => 'NL',
         ]);
         $response->assertStatus(403);
     }
 
     public function test_users_with_admin_role_can_update_other_member_roles()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $other_user = User::factory()->create();
 
@@ -115,13 +134,17 @@ class OrganizationTest extends TestCase
             ->actingAs($user)
             ->from(localized_route('memberships.edit', $membership))
             ->put(localized_route('memberships.update', $membership), [
-                'role' => 'admin'
+                'role' => 'admin',
             ]);
         $response->assertRedirect(localized_route('organizations.edit', $organization));
     }
 
     public function test_users_without_admin_role_can_not_update_member_roles()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
 
         $organization = Organization::factory()
@@ -137,7 +160,7 @@ class OrganizationTest extends TestCase
             ->actingAs($user)
             ->from(localized_route('memberships.edit', $membership))
             ->put(localized_route('memberships.update', $membership), [
-                'role' => 'admin'
+                'role' => 'admin',
             ]);
 
         $response->assertStatus(403);
@@ -145,6 +168,10 @@ class OrganizationTest extends TestCase
 
     public function test_only_administrator_can_not_downgrade_their_role()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $other_user = User::factory()->create();
 
@@ -162,7 +189,7 @@ class OrganizationTest extends TestCase
             ->actingAs($user)
             ->from(localized_route('memberships.edit', $membership))
             ->put(localized_route('memberships.update', $membership), [
-                'role' => 'member'
+                'role' => 'member',
             ]);
 
         $response->assertSessionHasErrors(['membership']);
@@ -171,6 +198,10 @@ class OrganizationTest extends TestCase
 
     public function test_users_with_admin_role_can_invite_members()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
 
         $organization = Organization::factory()
@@ -182,9 +213,9 @@ class OrganizationTest extends TestCase
             ->from(localized_route('organizations.edit', ['organization' => $organization]))
             ->post(localized_route('invitations.create'), [
                 'inviteable_id' => $organization->id,
-                'inviteable_type' => $organization->getModelClass(),
+                'inviteable_type' => get_class($organization),
                 'email' => 'newuser@here.com',
-                'role' => 'member'
+                'role' => 'member',
             ]);
 
         $response->assertRedirect(localized_route('organizations.edit', $organization));
@@ -192,6 +223,10 @@ class OrganizationTest extends TestCase
 
     public function test_users_without_admin_role_can_not_invite_members()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
 
         $organization = Organization::factory()
@@ -203,9 +238,9 @@ class OrganizationTest extends TestCase
             ->from(localized_route('organizations.edit', ['organization' => $organization]))
             ->post(localized_route('invitations.create'), [
                 'inviteable_id' => $organization->id,
-                'inviteable_type' => $organization->getModelClass(),
+                'inviteable_type' => get_class($organization),
                 'email' => 'newuser@here.com',
-                'role' => 'member'
+                'role' => 'member',
             ]);
 
         $response->assertStatus(403);
@@ -213,13 +248,17 @@ class OrganizationTest extends TestCase
 
     public function test_users_with_admin_role_can_cancel_invitations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $organization = Organization::factory()
             ->hasAttached($user, ['role' => 'admin'])
             ->create();
         $invitation = Invitation::factory()->create([
             'inviteable_id' => $organization->id,
-            'inviteable_type' => $organization->getModelClass(),
+            'inviteable_type' => get_class($organization),
             'email' => 'me@here.com',
         ]);
 
@@ -234,14 +273,18 @@ class OrganizationTest extends TestCase
 
     public function test_users_without_admin_role_can_not_cancel_invitations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $organization = Organization::factory()
             ->hasAttached($user, ['role' => 'member'])
             ->create();
         $invitation = Invitation::factory()->create([
             'inviteable_id' => $organization->id,
-            'inviteable_type' => $organization->getModelClass(),
-            'email' => 'me@here.com'
+            'inviteable_type' => get_class($organization),
+            'email' => 'me@here.com',
         ]);
 
         $response = $this
@@ -254,6 +297,10 @@ class OrganizationTest extends TestCase
 
     public function test_existing_members_cannot_be_invited()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $other_user = User::factory()->create();
 
@@ -267,9 +314,9 @@ class OrganizationTest extends TestCase
             ->from(localized_route('organizations.edit', ['organization' => $organization]))
             ->post(localized_route('invitations.create'), [
                 'inviteable_id' => $organization->id,
-                'inviteable_type' => $organization->getModelClass(),
+                'inviteable_type' => get_class($organization),
                 'email' => $other_user->email,
-                'role' => 'member'
+                'role' => 'member',
             ]);
 
         $response->assertSessionHasErrorsIn('inviteMember', ['email']);
@@ -278,12 +325,16 @@ class OrganizationTest extends TestCase
 
     public function test_invitation_can_be_accepted()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $organization = Organization::factory()->create();
         $invitation = Invitation::factory()->create([
             'inviteable_id' => $organization->id,
-            'inviteable_type' => $organization->getModelClass(),
-            'email' => $user->email
+            'inviteable_type' => get_class($organization),
+            'email' => $user->email,
         ]);
 
         $acceptUrl = URL::signedRoute('invitations.accept', ['invitation' => $invitation]);
@@ -296,11 +347,15 @@ class OrganizationTest extends TestCase
 
     public function test_invitation_cannot_be_accepted_unless_account_exists()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $organization = Organization::factory()->create();
         $invitation = Invitation::factory()->create([
             'inviteable_id' => $organization->id,
-            'inviteable_type' => $organization->getModelClass(),
-            'email' => 'me@here.com'
+            'inviteable_type' => get_class($organization),
+            'email' => 'me@here.com',
         ]);
 
         $acceptUrl = URL::signedRoute('invitations.accept', ['invitation' => $invitation]);
@@ -312,6 +367,10 @@ class OrganizationTest extends TestCase
 
     public function test_users_with_admin_role_can_remove_members()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $other_user = User::factory()->create();
 
@@ -336,6 +395,10 @@ class OrganizationTest extends TestCase
 
     public function test_users_without_admin_role_can_not_remove_members()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $other_user = User::factory()->create();
 
@@ -359,6 +422,10 @@ class OrganizationTest extends TestCase
 
     public function test_only_administrator_can_not_remove_themself()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
 
         $organization = Organization::factory()
@@ -381,37 +448,45 @@ class OrganizationTest extends TestCase
 
     public function test_users_with_admin_role_can_delete_organizations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $organization = Organization::factory()
             ->hasAttached($user, ['role' => 'admin'])
             ->create();
 
-        $response = $this->post('/en/login', [
+        $response = $this->post(localized_route('login'), [
             'email' => $user->email,
             'password' => 'password',
         ]);
 
         $response = $this->from(localized_route('organizations.edit', $organization))->delete(localized_route('organizations.destroy', $organization), [
-            'current_password' => 'password'
+            'current_password' => 'password',
         ]);
 
-        $response->assertRedirect('/en/dashboard');
+        $response->assertRedirect(localized_route('dashboard'));
     }
 
     public function test_users_with_admin_role_can_not_delete_organizations_with_wrong_password()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $organization = Organization::factory()
             ->hasAttached($user, ['role' => 'admin'])
             ->create();
 
-        $response = $this->post('/en/login', [
+        $response = $this->post(localized_route('login'), [
             'email' => $user->email,
             'password' => 'password',
         ]);
 
         $response = $this->from(localized_route('organizations.edit', $organization))->delete(localized_route('organizations.destroy', $organization), [
-            'current_password' => 'wrong_password'
+            'current_password' => 'wrong_password',
         ]);
 
         $response->assertSessionHasErrors();
@@ -420,18 +495,22 @@ class OrganizationTest extends TestCase
 
     public function test_users_without_admin_role_can_not_delete_organizations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $organization = Organization::factory()
             ->hasAttached($user, ['role' => 'member'])
             ->create();
 
-        $response = $this->post('/en/login', [
+        $response = $this->post(localized_route('login'), [
             'email' => $user->email,
             'password' => 'password',
         ]);
 
         $response = $this->from(localized_route('organizations.edit', $organization))->delete(localized_route('organizations.destroy', $organization), [
-            'current_password' => 'password'
+            'current_password' => 'password',
         ]);
 
         $response->assertStatus(403);
@@ -439,6 +518,10 @@ class OrganizationTest extends TestCase
 
     public function test_non_members_can_not_delete_organizations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $other_user = User::factory()->create();
 
@@ -450,13 +533,13 @@ class OrganizationTest extends TestCase
             ->hasAttached($other_user, ['role' => 'admin'])
             ->create();
 
-        $response = $this->post('/en/login', [
+        $response = $this->post(localized_route('login'), [
             'email' => $user->email,
             'password' => 'password',
         ]);
 
         $response = $this->from(localized_route('organizations.edit', $other_organization))->delete(localized_route('organizations.destroy', $other_organization), [
-            'current_password' => 'password'
+            'current_password' => 'password',
         ]);
 
         $response->assertStatus(403);
@@ -464,15 +547,19 @@ class OrganizationTest extends TestCase
 
     public function test_users_can_view_organizations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $user = User::factory()->create();
         $organization = Organization::factory()->create();
 
-        $response = $this->post('/en/login', [
+        $response = $this->post(localized_route('login'), [
             'email' => $user->email,
             'password' => 'password',
         ]);
 
-        $response = $this->get('/en/organizations');
+        $response = $this->get(localized_route('organizations.index'));
         $response->assertStatus(200);
 
         $response = $this->get(localized_route('organizations.show', $organization));
@@ -481,13 +568,16 @@ class OrganizationTest extends TestCase
 
     public function test_guests_can_not_view_organizations()
     {
+        if (! config('hearth.organizations.enabled')) {
+            return $this->markTestSkipped('Organization support is not enabled.');
+        }
+
         $organization = Organization::factory()->create();
 
-        $response = $this->get('/en/organizations');
-        $response->assertRedirect('/en/login');
+        $response = $this->get(localized_route('organizations.index'));
+        $response->assertRedirect(localized_route('login'));
 
         $response = $this->get(localized_route('organizations.show', $organization));
-        $response->assertRedirect('/en/login');
-
+        $response->assertRedirect(localized_route('login'));
     }
 }
